Morse algebra and types implemented
