+++ app/com/github/pedrovgs/dotto/actors/ShowMessageActor.scala
@@ -2,7 +2,9 @@ package com.github.pedrovgs.dotto.actors
 
 import akka.actor.{Actor, Props}
 import com.github.pedrovgs.dotto.actors.ShowMessageActor.ShowMessage
-import com.github.pedrovgs.dotto.core.algebra.morse._
+import com.github.pedrovgs.dotto.core.types.{LedInteraction, Low, High}
+import com.github.pedrovgs.dotto.morse.translator
+import translator._
 import com.github.pedrovgs.dotto.led._
 import com.pi4j.io.gpio._
 import play.api.Logger
