+++ app/com/github/pedrovgs/dotto/core/algebra/led.scala
@@ -1,5 +1,5 @@
 package com.github.pedrovgs.dotto.core.algebra
 
-class led {
+object led {
 
 }
