+++ app/com/github/pedrovgs/dotto/core/types/led.scala
@@ -1,5 +1,10 @@
 package com.github.pedrovgs.dotto.core.types
 
-class led {
+import scala.concurrent.duration.Duration
 
-}
+sealed trait PinState
+
+case object High extends PinState
+case object Low extends PinState
+
+case class LedInteraction(state: PinState, duration: Duration)
