+++ test/com/github/pedrovgs/dotto/led/LedControllerSpec.scala
@@ -1,6 +1,6 @@
 package com.github.pedrovgs.dotto.led
 
-import com.github.pedrovgs.dotto.core.types.{Dash, Dot, MorseSymbol, Space}
+import com.github.pedrovgs.dotto.core.types._
 import org.scalatest.FlatSpec
 
 import scala.concurrent.duration._
